+++ test/test_web.rb
@@ -9,7 +9,7 @@ describe Sidekiq::Web do
   include Rack::Test::Methods
 
   def app
-    Sidekiq::Web
+    @app ||= Sidekiq::Web.new
   end
 
   def job_params(job, score)
@@ -17,9 +17,8 @@ describe Sidekiq::Web do
   end
 
   before do
-    ENV["RACK_ENV"] = "test"
     Sidekiq.redis {|c| c.flushdb }
-    Sidekiq::Web.middlewares.clear
+    app.default_middleware.clear
   end
 
   class WebWorker
@@ -738,6 +737,7 @@ describe Sidekiq::Web do
     def app
       app = Sidekiq::Web.new
       app.use(Rack::Auth::Basic) { |user, pass| user == "a" && pass == "b" }
+      app.use(Rack::Session::Cookie, secret: SecureRandom.hex(32))
 
       app
     end
@@ -792,7 +792,9 @@ describe Sidekiq::Web do
     end
 
     def app
-      Sidekiq::Web.new
+      app = Sidekiq::Web.new
+      app.use Rack::Session::Cookie, secret: 'v3rys3cr31', host: 'nicehost.org'
+      app
     end
 
     it "allows afters to run" do